requirements.txt
Django==2.0
django-express==0.3.2
